Added some more utilities
