+++ packages/attendance/src/pages/reports/ReportDetail.js
@@ -14,7 +14,7 @@ import React, { useState, useEffect, Suspense } from "react";
 import { useTranslation } from "react-i18next";
 import CalendarBar from "components/composite/CalendarBar";
 import { GetAttendance } from "../../services/calls/registryCalls";
-import AttendanceComponent from "../../components/AttendanceComponent";
+import AttendanceComponent from "../../components/composite/AttendanceComponent";
 import ReportSummary from "../../components/ReportSummary";
 import { Link, useNavigate, useParams } from "react-router-dom";
