+++ packages/attendance/src/utils/functions/Constants.ts
@@ -1,0 +1,5 @@
+export const PRESENT = "Present";
+export const ABSENT = "Absent";
+export const UNMARKED = "Unmarked";
+export const MALE = "Male";
+export const FEMALE = "Female";
